Add @automattic/design-types to break circular dependency between data-stores and design-picker
