+++ packages/data-stores/src/onboard/reducer.ts
@@ -11,8 +11,7 @@ import type {
 import type { FeatureId } from '../shared-types';
 import type { GlobalStyles } from '../site';
 import type { DomainSuggestion } from '@automattic/api-core';
-// somewhat hacky, but resolves the circular dependency issue
-import type { Design, StyleVariation } from '@automattic/design-picker/src/types';
+import type { Design, StyleVariation } from '@automattic/design-types';
 import type { MinimalRequestCartProduct } from '@automattic/shopping-cart';
 import type { Reducer } from 'redux';
