+++ packages/data-stores/src/starter-designs-queries/types.ts
@@ -1,4 +1,4 @@
-import type { Category, Design } from '@automattic/design-picker/src/types';
+import type { Category, Design } from '@automattic/design-types';
 
 export interface StarterDesigns {
 	filters: { subject: Record< string, Category > };
