+++ packages/data-stores/src/starter-designs-queries/use-starter-designs-query.ts
@@ -10,7 +10,7 @@ import type {
 	StyleVariation,
 	PreviewData,
 	DesignType,
-} from '@automattic/design-picker/src/types';
+} from '@automattic/design-types';
 
 interface StarterDesignsQueryParams {
 	seed?: string;
