+++ packages/design-picker/src/utils/__tests__/designs.test.ts
@@ -1,6 +1,6 @@
 import { DEFAULT_VIEWPORT_HEIGHT } from '../../constants';
-import { Design, DesignPreviewOptions } from '../../types';
 import { getDesignPreviewUrl } from '../designs';
+import type { Design, DesignPreviewOptions } from '../../types';
 
 const mergeDesign = ( currentDesign, newDesign ) => ( {
 	...currentDesign,
