+++ packages/tsconfig.json
@@ -26,6 +26,7 @@
 		{ "path": "./composite-checkout" },
 		{ "path": "./create-calypso-config" },
 		{ "path": "./data-stores" },
+		{ "path": "./design-types" },
 		{ "path": "./design-picker" },
 		{ "path": "./design-preview" },
 		{ "path": "./domain-search" },
